Print all warnings to screen after finishing TE estimation
Test configuration structures for parameter combinations and return
warnings or comments on the TE estimation. For example, if the
Faes-method is used to deal with volume conduction, no test for
volume conduction can be performed when also calculating mutual
information. This is printed as a warning to the screen after the
estimation is done.